+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ### [0.2.5] - 2024-02-29
 
 - Improvement: the help info of `pull`.
+- Auth: Add `reuse` to `fav auth reuse` to reuse the old cookie.
 
 ## [0.2.4] - 2024-02-24